Make HMR updates work
This commit makes HMR updates work without having to import the entire
css inline. However, this has reduced the extension to only work for css
files (or files that compile to css files) so I have changed the name
accordingly.

How I've made HMR work is that when the server gets an HMR update, the
plugin scans for ?direct imports that have a corresponding ?css-url
import and triggers a module reload for the ?css-url modules. This meets
my use cases.