Append ?used to the asset id in build mode
Without having ?used as part of the asset query parameters the
vite:css-post plugin will only emit
    export default ''
as the entire module instead of including the CSS like we expect. This
leads to errors as '' is not a valid JSON string. The fix here is to
always add ?used as one of the id query parameters so that we get the
output we expect.